trying to get the VL6180x sensor to work. writes don't seem to be working still.

+++ src/org/usfirst/frc/team5461/sensors/VL6180xIdentification.java
@@ -7,6 +7,6 @@ public class VL6180xIdentification
   public int idModelRevMinor;
   public int idModuleRevMajor;
   public int idModuleRevMinor;
-  public short idDate;
-  public short idTime;
+  public int idDate;
+  public int idTime;
 }
